+++ keyledsd/include/keyledsd/effect/EffectService.h
@@ -48,6 +48,7 @@ public:
     const string_map &  configuration() const override;
     const std::string & getConfig(const std::string &) const override;
           bool          parseColor(const std::string &, RGBAColor *) const override;
+          bool          parseNumber(const std::string &, unsigned *) const override;
 
     RenderTarget *      createRenderTarget() override;
     void                destroyRenderTarget(RenderTarget *) override;